+++ SWLOR.Game.Server.Tests/Service/AchievementTrackingTests.cs
@@ -1,4 +1,3 @@
-using System.Reflection;
 using FluentAssertions;
 using NUnit.Framework;
 using SWLOR.Game.Server.Entity;
@@ -13,32 +12,8 @@ using Player = SWLOR.Game.Server.Entity.Player;
 
 namespace SWLOR.Game.Server.Tests.Service;
 
-public sealed class AchievementExpansionTests
+public sealed class AchievementTrackingTests
 {
-    [Test]
-    public void NewAchievements_AreActiveAndUseConsecutiveStableIds()
-    {
-        var achievements = Enum.GetValues<AchievementType>()
-            .Where(type => (int)type is >= 181 and <= 210)
-            .ToArray();
-
-        achievements.Select(type => (int)type).Should().Equal(Enumerable.Range(181, 30));
-        achievements.Should().HaveCount(30);
-
-        foreach (var achievement in achievements)
-        {
-            var detail = typeof(AchievementType)
-                .GetMember(achievement.ToString())
-                .Single()
-                .GetCustomAttribute<AchievementAttribute>();
-
-            detail.Should().NotBeNull();
-            detail!.IsActive.Should().BeTrue();
-            detail.Name.Should().NotBeNullOrWhiteSpace();
-            detail.Description.Should().NotBeNullOrWhiteSpace();
-        }
-    }
-
     [TestCase(FishingLocationType.ViscaraLake, PlanetType.Viscara)]
     [TestCase(FishingLocationType.MonCalaDacCitySurface, PlanetType.MonCala)]
     [TestCase(FishingLocationType.HutlarQionTundra, PlanetType.Hutlar)]
